수정:검색 페이지 필터 넣어서 서버로 요청 보내기, 메인 페이지 ajax 분리, header 수정 / 추가:SearchFilter, main.js, mainInit.js
