+++ src/main/java/com/kh/mng/search/model/dao/SearchDao.java
@@ -7,9 +7,9 @@ import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.stereotype.Repository;
 
 import com.kh.mng.common.model.vo.PageInfo;
-import com.kh.mng.location.model.vo.EnterGrade;
 import com.kh.mng.location.model.vo.Location;
 import com.kh.mng.location.model.vo.OperationTime;
+import com.kh.mng.search.model.dto.SearchFilter;
 
 @Repository
 public class SearchDao {
@@ -24,6 +24,14 @@ public class SearchDao {
 		RowBounds rowBounds = new RowBounds(offset, limit);
 		
 		return (ArrayList)sqlSession.selectList("searchMapper.selectSearchLocationList", keyword, rowBounds);
+	}
+	
+	public ArrayList<Location> selectFilterLocationList(SqlSessionTemplate sqlSession, SearchFilter sf, PageInfo pi){
+		int offset = (pi.getCurrentPage() - 1) * pi.getBoardLimit();
+		int limit = pi.getBoardLimit();
+		RowBounds rowBounds = new RowBounds(offset, limit);
+		System.out.println("여기 들어왔어?");
+		return (ArrayList)sqlSession.selectList("searchMapper.selectFilterLocationList", sf, rowBounds);
 	}
 	
 //	public ArrayList<EnterGrade> selectEnterGradeList(SqlSessionTemplate sqlSession, Location loc){
