+++ src/main/webapp/resources/js/ajax/ajax.js
@@ -1,0 +1,17 @@
+// 1. MainPage
+
+// 2. SearchPage
+
+// 3. DetailPage
+
+// 4. Community-Shorts
+
+// 5. Community-Board
+
+// 6. Login / Enroll
+
+// 7. MyPage
+
+// 8. BossPage
+
+// 9.
