+++ src/main/webapp/resources/js/search/searchAjax.js
@@ -9,30 +9,59 @@ function init(path, user){
     // document.querySelector('.like-btn').click();
 }
 
-function ajaxSelectLikeInfo(userNo, locationNo){
-    console.log('들어옴')
-    console.log(userNo)
-    if (userNo != ""){
-        console.log("여기는?")
-        ajaxGetData(contextPath + "/selectLikeInfo.pl", 
-        {
-            userNo : userNo,
-            locationNo : locationNo
-        },
-        function(result){console.log(result)});
+function searchFilter(contextPath, keyword){
+    console.log("들어옴~")
+    let pets = document.querySelectorAll('.filter-pet');
+    let locs = document.querySelectorAll('.filter-location');
+    let order = document.querySelectorAll('[name="order"]');
+    // let order = document.querySelector('.order-by-list');
+    // let cpage;
+
+    let petList = [];
+    for (let p of pets){
+        if (p.checked){
+            petList.push(p.value)
+        }
     }
+
+    let locList = [];
+    for (let l of locs){
+        if (l.checked){
+            locList.push(l.value)
+        }
+    }
+
+    let cpage = 1;
+
+    getLocationData({
+        keyword : keyword,
+        petList : petList.toString(),
+        locList : locList.toString(),
+        order : order.value,
+        cpage : cpage
+    }, function(res){
+        console.log(res)
+    })
 }
 
-// function ajaxSelectLocation(keyword, cpage){
-//     ajaxGetData(contextPath + "/searchKeyword.pl", 
-//     {keyword: keyword},
-
-//     function(result){drawPlaceTop(result)});
-// }
+function getLocationData(data, callback){
+    $.ajax({
+        url: contextPath + "/searchPage.pl",
+        data: data,
+        success: function (result) {
+            callback(result)
+        },
+        error: function () {
+            console.log("정보를 불러오는데 실패 했습니다.");
+        }
+    })
+}
 
 function drawSearchPage(locationList){
+    console.log("성공했어요~")
     let searchResultBox = document.querySelector("#search-result-box");
 
+    searchResultBox.innerHTML = "";
     // for(let loc of locationList){
     //     let searchContentBox = document.createElement('div');
     //     searchContentBox.classList.add('search-content-box', 'gray-round-box');
@@ -63,20 +92,21 @@ function drawSearchPage(locationList){
     
 }
 
+// 좋아요
+// function ajaxSelectLikeInfo(userNo, locationNo){
+//     console.log('들어옴')
+//     console.log(userNo)
+//     if (userNo != ""){
+//         console.log("여기는?")
+//         ajaxGetData(contextPath + "/selectLikeInfo.pl", 
+//         {
+//             userNo : userNo,
+//             locationNo : locationNo
+//         },
+//         function(result){console.log(result)});
+//     }
+// }
+
 // function drawSearchPlace(){
 
 // }
-
-// ajax 기본 함수
-function ajaxGetData(url, data, callback){
-    $.ajax({
-        url: url,
-        data: data,
-        success: function (result) {
-            callback(result)
-        },
-        error: function () {
-            console.log("정보를 불러오는데 실패 했습니다.");
-        }
-    })
-}
